+++ src/anim/ui/ui_button.cpp
@@ -39,10 +39,10 @@ public:
 
 
 
-button::button( animation *Anim, const element &Base,
+button::button( animation *Anim, const clickable &Base,
                 const callback &Callback,
                 std::shared_ptr<material> Material ) :
-    element(Base), Callback(Callback), Material(Material) {
+    clickable(Base), Callback(Callback), Material(Material) {
   Primitive.Set(Anim);
   Primitive.Set(std::shared_ptr<geometry>(
     new geometry(1, new button_vertex[1]{ button_vertex(BoundBox[0] * 2.f + point<float>(-1.f, -1.f),