+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 All notable changes to this project will be documented in this file.
 
 
-## [1.0.1] - 2020-12-21
+## [1.2.0] - 2020-12-21
 ### Added:
-- 'hideable' ui element
+- 'ui::clickable' ui element
+
+### Changed:
+- Remove all logic from ui::element, move it to ui::clickable
+
+
+## [1.1.0] - 2020-12-21
+### Added:
+- 'ui::hideable' ui element
 
 
 ## [1.0.0] - 2020-12-20